+++ CHANGELOG.md
@@ -6,6 +6,13 @@
   - Example: 10.2.1.4 is the 5th version that supports khiops 10.2.1.
 - Internals: Changes in *Internals* sections are unlikely to be of interest for data scientists.
 
+## Unreleased
+
+### Added
+
+- (General) `visualize_report` helper function to open reports with the Khiops Visualization and
+  Khiops Co-Visualization app.
+
 ## 10.2.3.1 - 2024-11-27
 
 ### Added